refactor(maps): move map components into nested directories
