+++ lib/maps/ChoroplethMap/ChoroplethMap.vue
@@ -28,9 +28,9 @@ import {
 
 import { ParentKey } from '@/keys'
 import { MapTransform, ParentMap } from '@/types'
-import config from '../config'
+import config from '@/config'
 import { getChartProps, useChart } from '@/composables/useChart'
-import ScaleLegend from '@/components/ScaleLegend.vue'
+import ScaleLegend from '@/components/Legend/ScaleLegend.vue'
 
 defineOptions({
   name: 'ChoroplethMap'
