+++ lib/maps/ChoroplethMap/ChoroplethMapAnnotation.vue
