+++ lib/maps/index.js
@@ -1,3 +1,3 @@
-export { default as ChoroplethMap } from './ChoroplethMap.vue'
-export { default as ChoroplethMapAnnotation } from './ChoroplethMapAnnotation.vue'
-export { default as SymbolMap } from './SymbolMap.vue'
+export { default as ChoroplethMap } from './ChoroplethMap/ChoroplethMap.vue'
+export { default as ChoroplethMapAnnotation } from './ChoroplethMap/ChoroplethMapAnnotation.vue'
+export { default as SymbolMap } from './SymbolMap/SymbolMap.vue'
